+++ mobile/lib/services/sync/local_sync_service.dart
@@ -7,8 +7,8 @@ import 'package:photo_manager/photo_manager.dart';
 import 'package:photos/core/configuration.dart';
 import "package:photos/core/errors.dart";
 import 'package:photos/core/event_bus.dart';
+import "package:photos/db/common/conflict_algo.dart";
 import 'package:photos/db/device_files_db.dart';
-import "package:photos/db/enum/conflict_algo.dart";
 import 'package:photos/db/file_updation_db.dart';
 import 'package:photos/db/files_db.dart';
 import 'package:photos/events/backup_folders_updated_event.dart';
